Updateted some tasks

+++ 02.Data-Types-and-Variables/14.Print-The-ASCII-Table/PrintTheASCIITable.cs
@@ -1,7 +1,11 @@
 ﻿using System;
 
+
 /// <summary>
-/// 
+/// Find online more information about ASCII (American Standard Code for Information Interchange) 
+/// and write a program that prints the visible characters of the ASCII table on the console 
+/// (characters from 33 to 126 including).
+/// Note: You may need to use for-loops(learn in Internet how).
 /// </summary>
 class PrintTheASCIITable
 {
@@ -10,7 +14,8 @@ class PrintTheASCIITable
     /// </summary>
     static void Main()
     {
-        for (int i = 33; i < 127; i++)
+        
+        for (byte i = 33; i < 127; i++)
         {
             Console.WriteLine((char)i);
         }
